Converted concept map component to ts. #2258
