+++ src/main/webapp/wise5/components/conceptMap/conceptMapAuthoringComponentModule.ts
@@ -3,16 +3,16 @@
 import ConceptMapService from './conceptMapService';
 import ConceptMapController from './conceptMapController';
 import ConceptMapAuthoringController from './conceptMapAuthoringController';
+import * as angular from 'angular';
 
-const conceptMapAuthoringComponentModule = angular.module('conceptMapAuthoringComponentModule', [
-    'pascalprecht.translate'
-  ])
+const conceptMapAuthoringComponentModule = angular
+  .module('conceptMapAuthoringComponentModule', ['pascalprecht.translate'])
   .service('ConceptMapService', ConceptMapService)
   .controller('ConceptMapController', ConceptMapController)
   .controller('ConceptMapAuthoringController', ConceptMapAuthoringController)
   .config([
     '$translatePartialLoaderProvider',
-    ($translatePartialLoaderProvider) => {
+    $translatePartialLoaderProvider => {
       $translatePartialLoaderProvider.addPart('components/conceptMap/i18n');
     }
   ]);
